% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freetext_functions.R
\name{extract_words}
\alias{extract_words}
\title{Extract words strings}
\usage{
extract_words(free_text_vector)
}
\arguments{
\item{free_text_vector}{a character vector of free text}
}
\value{
list with one vector per value in the input vector. Each element in the vector is a word with the first letter
in capital.
}
\description{
Extract words strings
}
\details{
splits words on spaces. removes punctuation and white space.
}
