% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freetext_functions.R
\name{word_frequency}
\alias{word_frequency}
\title{Counts words in free text}
\usage{
word_frequency(free_text_vector)
}
\arguments{
\item{free_text_vector}{a character vector of free text}
}
\value{
A dataframe with 2 columns, "word": the identified word; "count": how many time the word appeared.
}
\description{
Counts words in free text
}
\details{
Counts words identified with `extract_word()`
}
